requirements.txt
flask-cors

Flask==1.0.2
redis==2.10.6
gunicorn==19.9.0
gevent==1.5a2
gunicorn==19.9.0
requests==2.21.0

# kafka
kafka-python==1.4.6

# Elastic logstash
elastic-apm[flask]==4.1.0
python3-logstash==0.4.80
python-json-logger==0.1.11
psutil==5.6.7

# unittest command
green==2.16.1

pymongo==3.7.2